Add : 삭제 프롬프트 값이 맞으면 취소시에도 삭제되는 버그 수정

+++ src/stores/home.js
@@ -39,14 +39,17 @@ export const useHomeStore = defineStore('home', {
         // 데이터 삭제
         deleteEmotion(emotionPassword, emotionId) {
             let pass = prompt('삭제를 위해 등록 시 입력한 암호를 입력해주세요.');
-            if (emotionPassword == pass) {
-                let emoRef = doc(emoColRef, emotionId);
-                deleteDoc(emoRef).then(() => {
-                    alert("감정이 삭제되었습니다");
-                    window.location.reload();
-                });
-            } else {
-                alert("암호가 맞지 않습니다.");
+            console.log(pass);
+            if (pass !== null) {
+                if (emotionPassword == pass) {
+                    let emoRef = doc(emoColRef, emotionId);
+                    deleteDoc(emoRef).then(() => {
+                        alert("감정이 삭제되었습니다");
+                        window.location.reload();
+                    });
+                } else {
+                    alert("암호가 맞지 않습니다.");
+                }
             }
         },
     },
